+++ src/api/my-info/index.ts
@@ -1,0 +1,2 @@
+export * from './content';
+export * from './member';
