+++ src/types/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './login';
 export * from './signUp';
+export * from './token';
